feat: rename function to useFindSessionQuery
